requirements.txt
aws-cdk.core>=1.80.0
aws-cdk.aws_lambda>=1.80.0
aws-cdk.aws_ecr>=1.80.0
aws-cdk.aws_apigatewayv2>=1.80.0
aws-cdk.aws_apigatewayv2-integrations>=1.80.0